fix: move networking action visibility helper out of component module
